Fix for breaking change in GHC 7.10.1
With GHC 7.10.1 or later and without the language extension FlexibleContexts, building will give errors like,
"Ambiguous type variable ‘x’ arising from a use of ‘y’ prevents the constraint ‘(y x)’ from being solved."

From
[GHC 7.10.1 release notes](https://downloads.haskell.org/~ghc/7.10.1-rc1/docs/html/users_guide/release-7-10-1.html)
the change is

```
GHC now checks that all the language extensions required for the inferred type signatures are explicitly enabled. This means that if any of the type signatures inferred in your program requires some language extension you will need to enable it. The motivation is that adding a missing type signature inferred by GHC should yield a program that typechecks. Previously this was not the case.

This is a breaking change. Code that used to compile in the past might fail with an error message requiring some particular language extension (most likely -XTypeFamilies, -XGADTs or -XFlexibleContexts).
```